requirements.txt
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
pymongo==4.6.3
dnspython==2.6.1
python-dotenv==1.0.1